Initial pong animation
